group controller bis

+++ app/src/main/kotlin/me/jvegaf/user/domain/UserRepository.kt
@@ -4,6 +4,6 @@ import io.micronaut.data.annotation.Repository
 import io.micronaut.data.jpa.repository.JpaRepository
 
 @Repository
-interface UserRepository : JpaRepository<User, Long> {
+interface UserRepository : JpaRepository<User, Int> {
     fun findByName(name: String): User?
 }
